+++ parsing/padding.go
@@ -12,7 +12,7 @@ func Pad(unpaddedData []byte) ([]fr32.Fr32, error) {
 		return nil, errors.New("empty input")
 	}
 	// Compute amount of Fr32 elements in the result
-	chunkCount := util.Ceil(len(unpaddedData)*8, fr32.BitsNeeded)
+	chunkCount := util.Ceil(uint(len(unpaddedData)*8), fr32.BitsNeeded)
 	paddedData := make([]fr32.Fr32, chunkCount)
 	bitIdx := 0
 	for i := 0; i < chunkCount; i++ {
@@ -62,7 +62,7 @@ func Unpad(paddedData []fr32.Fr32) ([]byte, error) {
 		return nil, errors.New("empty input")
 	}
 	// Compute amount of bytes in the result
-	bytes := util.Ceil(len(paddedData)*fr32.BitsNeeded, 8)
+	bytes := util.Ceil(uint(len(paddedData)*fr32.BitsNeeded), 8)
 	unpaddedData := make([]byte, bytes)
 	bitIdx := 0
 	for i := 0; i < len(paddedData); i++ {
